refactor: implement DIP for phone/landline validators.
