+++ src/Utils/CountryLandlineValidator/INLandlineValidator.php
@@ -2,6 +2,8 @@
 
 namespace Milwad\LaravelValidate\Utils\CountryLandlineValidator;
 
+use Milwad\LaravelValidate\Utils\CountryLandlineValidator;
+
 class INLandlineValidator implements CountryLandlineValidator
 {
     /**
