+++ src/Utils/CountryPhoneValidator/CMPhoneValidator.php
@@ -2,6 +2,8 @@
 
 namespace Milwad\LaravelValidate\Utils\CountryPhoneValidator;
 
+use Milwad\LaravelValidate\Utils\CountryPhoneValidator;
+
 class CMPhoneValidator implements CountryPhoneValidator
 {
     public function validate($value): bool
