+++ src/Utils/CountryPhoneValidator/DEPhoneValidator.php
@@ -2,6 +2,8 @@
 
 namespace Milwad\LaravelValidate\Utils\CountryPhoneValidator;
 
+use Milwad\LaravelValidate\Utils\CountryPhoneValidator;
+
 class DEPhoneValidator implements CountryPhoneValidator
 {
     /**
